timestamp로 브랜치 생성 순 확인

+++ cli/lib/execSync.mjs
@@ -3,6 +3,7 @@ import {getResponse} from "./promft.mjs";
 import {isUsableRepoName, mkRepo} from "./api.mjs";
 import chalk from "chalk";
 import fs from "fs";
+import path from "path";
 
 /** git repository 여부 **/
 export async function existsGitDirectory(){
@@ -37,34 +38,62 @@ export async function getLocalBranches() {
     return branches;
 }
 
-/** 브랜치의 첫 커밋 매칭 **/
-export async function getFirstCommitOfBranch(branch) {
+/** 브랜치 생성 순 반환 **/
+export async function getBranchCreationTimes() {
+    const dir = ".git/logs/refs/heads";
+    const branches = fs.readdirSync(dir);
+    const result = [];
+
+    for (const branch of branches) {
+        const content = fs.readFileSync(path.join(dir, branch), "utf-8");
+        const firstLine = content.split("\n")[0];
+        const parts = firstLine.split(" ");
+
+        const fromHash = parts[0];
+        const toHash = parts[1];
+        const timestamp = parseInt(parts[4], 10);
+        const event = firstLine.split('\t')[1];
+
+        result.push({ branch, timestamp, fromHash, toHash, event });
+    }
+
+    result.sort((a, b) => a.timestamp - b.timestamp);
+    return result;
+}
+
+/** 브랜치의 첫 이벤트 **/
+function getFirstCommitOfBranch(branch) {
 
     if (!fs.existsSync(`.git/logs/refs/heads/${branch}`)) return null;
-    const firstHistory = execSync(`head -n 1 .git/logs/refs/heads/${branch}`).toString().trim();;
+    const firstHistory = execSync(`head -n 1 .git/logs/refs/heads/${branch}`).toString().trim();
 
     // from to author <email> UNIXtimeStamp timeZone	clone: from https://github.com/teamProject-Y/DiFF.git
     // from to author <email> UNIXtimeStamp timeZone	branch: Created from HEAD
     console.log(firstHistory);
     const info = firstHistory.split('\t')[0].split(' ', 6);
     console.log(info[0]);
-    console.log(chalk.bgCyanBright(info[1]));
+    console.log(chalk.bgCyanBright(chalk.black(info[1])));
     console.log(info[2]);
     console.log(info[3]);
-    console.log(info[4]);
+    console.log(chalk.bgCyanBright(chalk.black(info[4])));
     console.log(info[5]);
 
     const commitInfo = firstHistory.split('\t')[1];
-    console.log(commitInfo);
+    console.log(chalk.bgCyanBright(chalk.black("git event: ", commitInfo)));
 
     let type;
-    if(commitInfo.startsWith('commit (initial):' || 'clone:')) { // commit ? commit (initial): ?
+    if(commitInfo.startsWith('commit' || 'clone:')) {
         type = "default";
     } else {
         type = "branch";
     }
 
-    return { branch, checksum: info[1], type, logInfo: commitInfo };
+    return { branch, checksum: info[1], UNIXtimeStamp: info[4], type, logInfo: commitInfo };
+}
+
+/** 요청 브랜치 마지막 체크섬 **/
+export function getLastChecksum(branch) {
+    return execSync(`git rev-parse ${branch}`).toString().trim();
 }
 
 /** .DiFF 디렉토리 만들기 **/
@@ -81,8 +110,8 @@ export async function DiFFinit(memberId, branch) {
         usable = await isUsableRepoName(memberId, repoName);
     }
 
-    console.log(chalk.bgYellow("repoName: " + repoName));
-    console.log(chalk.bgYellow("usable: " + usable));
+    console.log(chalk.bgCyanBright(chalk.black("repoName: " + repoName)));
+    console.log(chalk.bgCyanBright(chalk.black("usable: " + usable)));
 
     // git log --reverse --pretty=oneline ${nowBranch} ^${default} | head -n 1
     // head -n 1 .git/logs/refs/heads/yunzivv
@@ -93,14 +122,14 @@ export async function DiFFinit(memberId, branch) {
 
     const checksum = firstCommit.split(' ')[0];
     const commitMessage = firstCommit.split(' ').slice(1).join(' ');
-    console.log(chalk.bgYellow("first commit: ", firstCommit));
+    console.log(chalk.bgCyanBright(chalk.black("first commit: ", firstCommit)));
 
     // 서버에 리포지토리 DB 데이터 생성 요청
     let mkRepoAndGetId = await mkRepo(memberId, repoName, checksum);
     if(mkRepoAndGetId === null){
         return null;
     }
-    console.log(chalk.bgYellow("repoID: ", mkRepoAndGetId))
+    console.log(chalk.bgCyanBright(chalk.black("repoID: ", mkRepoAndGetId)));
 
     // .DiFF 생성
 
@@ -110,21 +139,20 @@ export async function DiFFinit(memberId, branch) {
     return checksum;
 }
 
-
 /** .DiFF 디렉토리에서 브랜치 내용 읽기 **/
 export async function gg(){
 
-    const isGitDirectory = execSync('[ -d .git ] && echo true || echo false').toString().trim();
+    const isDiFFdirectory = execSync('[ -d .DiFF ] && echo true || echo false').toString().trim();
 
-    if(isGitDirectory === 'false') {
-        console.log('fatal: not a git repository (or any of the parent directories): .git');
+    if(isDiFFdirectory === 'false') {
+        console.log('fatal: not a DiFF repository: .DiFF');
         return null;
     }
-    return isGitDirectory;
+    return isDiFFdirectory;
 }
 
 /** DiFF 디렉토리 만들기 **/
-export async function mkDiFFdirectory(){
+export async function mkDiFFdirectory(branches){
 
     if (fs.existsSync(".DiFF")) return null;
 
@@ -132,29 +160,38 @@ export async function mkDiFFdirectory(){
     const config = execSync("touch .DiFF/config");
     const meta = execSync("touch .DiFF/meta");
     const branchesDir = execSync("mkdir .DiFF/branches");
+    execSync("cd .DiFF/branches");
+    for(let branch in branches){
+
+    }
 
     return null;
 }
 
 /** zip 파일 **/
 export function mkZip(branch) {
-
     try {
-        const target = execSync(`[ -d target ] && echo true || echo false`);
-        execSync(`
-            git archive --format=zip --output=withoutTarget.zip ${branch} &&
-            rm -rf tempdir &&
-            mkdir tempdir &&
-            unzip withoutTarget.zip -d tempdir &&
-            cp -r target tempdir/ &&
-            cd tempdir && zip -r ../difftest.zip . &&
-            cd .. &&
-            rm withoutTarget.zip &&
-            rm -rf tempdir
-        `);
-        console.log("zip su");
+        const hasTarget = execSync(`[ -d target ] && echo true || echo false`).toString().trim();
 
+        if (hasTarget === "true") {
+            execSync(`
+                git archive --format=zip --output=withoutTarget.zip ${branch} &&
+                rm -rf tempdir && rm -rf difftest.zip &&
+                mkdir tempdir &&
+                unzip withoutTarget.zip -d tempdir &&
+                cp -r target tempdir/ &&
+                cd tempdir && zip -r ../difftest.zip . &&
+                cd .. &&
+                rm withoutTarget.zip &&
+                rm -rf tempdir
+            `);
+        } else {
+            execSync(`git archive --format=zip --output=difftest.zip ${branch}`);
+        }
+
+        console.log(chalk.bgCyanBright(chalk.black("zip success")));
         execSync(`curl -X POST -F "file=@difftest.zip" http://localhost:8080/upload`);
+        execSync(`rm -f difftest.zip`);
         return true;
 
     } catch (err) {
@@ -163,28 +200,28 @@ export function mkZip(branch) {
     }
 }
 
-export function getLastChecksum(branch) {
-    return execSync(`git rev-parse ${branch}`).toString().trim();
-}
-
+/** 첫 체크섬과 마지막 체크섬 사이의 diff **/
 export function getDiFF(firstChecksum, lastChecksum) {
-    const command = `git diff -W ${firstChecksum} ${lastChecksum} | grep -E "^[+-]|^@@"`.trim().replace(/\n/g, ' ');
-    return execSync(`sh -c '${command}'`).toString();
+
+    const command = `git diff -W f2b1257e9dfa51eea690b1872b99f8ffc21ba731 ${lastChecksum} | grep -E "^[+-]|^@@"`.trim().replace(/\n/g, ' ');
+    return execSync(`sh -c '${command}'`, {
+        maxBuffer: 1024 * 1024 * 50 // 50MB 까지 버퍼 허용
+    }).toString();
 
 }
 
-
+/** 로딩 애니메이션 **/
 function startAsciiAnimation() {
     const frames = [ `wating`, `...frame2...`, `...frame3...`, `...frame4...` ]; // 위 내용 넣기
     let index = 0;
 
     console.log("start 압축")
     const interval = setInterval(() => {
-        process.stdout.write('\x1Bc'); // clear terminal
+        process.stdout.write('\x1Bc');
         console.log(frames[index % frames.length]);
         index++;
     }, 2000);
 
-    return interval; // 나중에 clearInterval() 호출할 수 있도록
+    return interval;
 }
